--- src/Components/Home/home.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import Sidebar from '../SideBar/sidebar'
-import Cardcomp from './Cardcomp/cardcomp';
-function Home() {
-  return (
-    <>
-        <Sidebar />
-        <Cardcomp />
-    </>
-  )
-}
-
-export default Home;
