+++ src/Routers/Router.js
@@ -1,11 +1,14 @@
 // import { Category, Home, Login } from '@mui/icons-material'
 import React from 'react';
 import { BrowserRouter as Router , Routes , Route } from 'react-router-dom';
-import Home from '../Components/Home/home';
+import Home from '../pages/Home/home';
 import Login from '../pages/Login/Login';
-import  Recipe from '../Components/Home/Cardcomp/AddRecipeComp/recipe';
-import Category from '../Components/Home/Cardcomp/Category/category';
-import Submit from '../Components/Home/Cardcomp/Submit/Submit';
+
+import Category from '../pages/Categories/category';
+import Submit from '../pages/Suggestions/Submit';
+import Recipe from '../pages/AddRecipe/Recipe';
+
+
 
 
 
@@ -15,7 +18,7 @@ function AppRouter() {
     <Router>
         <Routes>
             <Route path='/' element={<Login />} />
-            <Route path='/home' element={<Home />} />
+            <Route path='/home' element={<Home/>} />
             <Route path='/addRecipe' element={<Recipe/>}/>
             <Route path='/submit' element={<Submit/>}/>
             <Route path='/category' element={<Category/>}/>
