+++ src/pages/Home/home.jsx
@@ -1,0 +1,15 @@
+import React from 'react'
+
+
+import Cardcomp from '../CardComponents/cardcomp';
+import SideBar from '../../Layout/SideBar/sidebar';
+function Home() {
+  return (
+    <>
+        <SideBar />
+        <Cardcomp />
+    </>
+  )
+}
+
+export default Home;
